refactor: move JAX equivalence tests to mrvi_torch, add @pytest.mark.jax
- Move test_mrvi_full_model_equivalence.py → mrvi_torch/test_torchmrvi_jax_equivalence.py
- Add @pytest.mark.jax so tests only run with --jax flag
- Use pytest.importorskip("jax") instead of top-level JAX imports
- Remove if __name__ == "__main__" block
